% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commutation_matrix.R
\name{commutation_matrix}
\alias{commutation_matrix}
\title{Commutation matrix K}
\usage{
commutation_matrix(r, c = r, checks = TRUE)
}
\arguments{
\item{r}{positive integer for row dimension.}

\item{c}{positive integer for column dimension.}

\item{checks}{logical, apply checks to arguments.}
}
\value{
matrix with \eqn{rc} rows and \eqn{rc} columns.
}
\description{
Let
\eqn{A} be a \eqn{r} by \eqn{c} matrix, then
\eqn{K} defines a commutation matrix such that
\eqn{vec(A) = K vec(A')} where prime denotes transpose.
}
\examples{
r = 2
c = 2
K <- commutation_matrix(r, c)
##      [,1] [,2] [,3] [,4]
## [1,]    1    0    0    0
## [2,]    0    0    1    0
## [3,]    0    1    0    0
## [4,]    0    0    0    1

# verifying vec(A) = K vec(A') for r=c=2 case
A <- matrix(1:(r*c), r, c)
vecA <- c(A)
vecAt <- c(t(A))
print(vecA)
## [1] 1 2 3 4
print(c(K \%*\% vecAt))
## [1] 1 2 3 4
}
\keyword{function}
\keyword{matrix}
