% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_deriv.R
\name{D_norm_2nd}
\alias{D_norm_2nd}
\title{Second Order Derivative of Euclidean norm}
\usage{
D_norm_2nd(x)
}
\arguments{
\item{x}{numeric vector.}
}
\value{
numeric vector of length x.
}
\description{
Second Order Derivative of Euclidean norm of vector \code{x} with respect
to \code{x}.
}
\examples{
library(numDeriv)
func <- function(x) norm(x, type="2")
x <- 1:2
print(numDeriv::hessian(func, x))
##           [,1]        [,2]
##[1,]  0.3577709 -0.17888544
##[2,] -0.1788854  0.08944272
print(D_norm_2nd(x))
##           [,1]        [,2]
##[1,]  0.3577709 -0.17888544
##[2,] -0.1788854  0.08944272
}
\seealso{
\link{D_norm_1st}, \link{D_norm_3rd}, \link{D_norm_4th}
}
