% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_deriv.R
\name{D_norm_3rd}
\alias{D_norm_3rd}
\title{Third Order Derivative of Euclidean norm}
\usage{
D_norm_3rd(x)
}
\arguments{
\item{x}{numeric vector.}
}
\value{
numeric vector of length x.
}
\description{
Third Order Derivative of Euclidean norm of vector \code{x} with respect
to \code{x}.
}
\examples{
func <- function(x) norm(x, type="2")
x <- 1:2
print(vector_deriv_3rd(func, x))
##             [,1]        [,2]
## [1,] -0.21466254 -0.07155418
## [2,] -0.07155418  0.12521981
## [3,] -0.07155418  0.12521981
## [4,]  0.12521981 -0.10733128
print(D_norm_3rd(x))
##             [,1]        [,2]
## [1,] -0.21466253 -0.07155418
## [2,] -0.07155418  0.12521981
## [3,] -0.07155418  0.12521981
## [4,]  0.12521981 -0.10733126
}
\seealso{
\link{D_norm_1st}, \link{D_norm_2nd}, \link{D_norm_4th}
}
