% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elimination_matrix.R
\name{elimination_matrix}
\alias{elimination_matrix}
\title{Elimination matrix L_m}
\usage{
elimination_matrix(m)
}
\arguments{
\item{m}{dimension parameter of \eqn{A}.}
}
\value{
matrix with \eqn{m(m+1)/2} rows and \eqn{m^2} columns.
}
\description{
Let
\eqn{A} be a \eqn{m} by \eqn{m} symmetric matrix, then
\eqn{L_m} defines an elimination matrix such that
\eqn{vech(A) = L_m vec(A)}
}
\examples{
elimination_matrix(2)
##      [,1] [,2] [,3] [,4]
## [1,]    1    0    0    0
## [2,]    0    1    0    0
## [3,]    0    0    0    1

# verifying vech(A) = L_m vec(A) for m=2 case
m <- 2
A <- matrix(1:m^2, m, m)
print(vech(A))
## [1] 1 2 4
print(c(elimination_matrix(m) \%*\% c(A)))
## [1] 1 2 4
}
\seealso{
\link{duplication_matrix}
}
\keyword{function}
\keyword{matrix}
