% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_invertible.R
\name{is_invertible}
\alias{is_invertible}
\title{Check if Matrix is invertible}
\usage{
is_invertible(X, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{X}{numeric square matrix.}

\item{tol}{
A relative tolerance to detect zero singular values.
}
}
\value{
TRUE if numerically invertible, FALSE otherwise.
}
\description{
Checks if the matrix is numerically invertible.
}
\examples{
X <- matrix(c(1,1,0,0), 2, 2)
is_invertible(X)

Y <- diag(2)
is_invertible(Y)
}
