% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_deriv.R
\name{matrix_deriv}
\alias{matrix_deriv}
\title{1st order matrix derivative of a matrix function}
\usage{
matrix_deriv(
  func,
  X,
  transpose_deriv_operator = FALSE,
  method = "Richardson",
  method.args = list(),
  ...
)
}
\arguments{
\item{func}{matrix function wrt matrix X.}

\item{X}{numeric matrix.}

\item{transpose_deriv_operator}{logical; apply transpose to the
derivative operator \eqn{d/dX}.}

\item{method}{currently only \code{"Richardson"} is implemented for the approximation.}

\item{method.args}{arguments passed to method. Arguments not specified 
       remain with their default values as specified in details}

\item{...}{an additional arguments passed to \code{func}.
          WARNING: None of these should have names matching other arguments of this function.}
}
\value{
matrix with \eqn{rp} rows and \eqn{cq} columns.
}
\description{
Let
\eqn{X} be a matrix of size \eqn{r} by \eqn{c},
\eqn{F} be a matrix function (wrt \eqn{X}) of size \eqn{p} by \eqn{q},
then \eqn{dF/dX} is the matrix derivative of F of size \eqn{rp} by \eqn{cq}
}
\examples{
func_11 <- function(X) X[1,1] * matrix(1:4,2,2)
func_12 <- function(X) X[1,2] * matrix(1:4,2,2)
func_21 <- function(X) X[2,1] * matrix(1:4,2,2)
func_22 <- function(X) X[2,2] * matrix(1:4,2,2)

X <- matrix(1, 2, 2)
matrix_deriv(func_11, X)
##      [,1] [,2] [,3] [,4]
## [1,]    1    3    0    0
## [2,]    2    4    0    0
## [3,]    0    0    0    0
## [4,]    0    0    0    0
matrix_deriv(func_12, X)
##      [,1] [,2] [,3] [,4]
## [1,]    0    0    1    3
## [2,]    0    0    2    4
## [3,]    0    0    0    0
## [4,]    0    0    0    0
matrix_deriv(func_21, X)
##      [,1] [,2] [,3] [,4]
## [1,]    0    0    0    0
## [2,]    0    0    0    0
## [3,]    1    3    0    0
## [4,]    2    4    0    0
matrix_deriv(func_22, X)
##      [,1] [,2] [,3] [,4]
## [1,]    0    0    0    0
## [2,]    0    0    0    0
## [3,]    0    0    1    3
## [4,]    0    0    2    4
}
\seealso{
\link{vector_deriv_3rd}, \link{vector_deriv_4th}, \link[numDeriv:grad]{numDeriv::grad}, \link[numDeriv:hessian]{numDeriv::hessian}
}
