% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spherical_moments.R
\name{spherical_moment_0th}
\alias{spherical_moment_0th}
\title{Zeroth order spherical moment}
\usage{
spherical_moment_0th(x, r = 1)
}
\arguments{
\item{x}{numeric vector, exponent in the exponential function of the integrand.}

\item{r}{positive numeric, radius of the spherical integral.}
}
\value{
Scalar.
}
\description{
Spherical integral of \eqn{exp(s'x)} where prime denotes transpose.
}
\examples{
spherical_moment_0th(1:3, 2)
}
\seealso{
\link{spherical_moment_1st}, \link{spherical_moment_2nd}, \link{spherical_moment_3rd}, \link{spherical_moment_4th}
}
