% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spherical_moments.R
\name{spherical_moment_2nd}
\alias{spherical_moment_2nd}
\title{Second order spherical moments}
\usage{
spherical_moment_2nd(x, r = 1)
}
\arguments{
\item{x}{numeric vector, exponent in the exponential function of the integrand.}

\item{r}{positive numeric, radius of the spherical integral.}
}
\value{
Numeric matrix
}
\description{
Second order derivative of Sperical integral of \eqn{exp(s'x)} where prime denotes transpose.
}
\examples{
spherical_moment_2nd(1:3, 2)
}
\seealso{
\link{spherical_moment_0th}, \link{spherical_moment_1st}, \link{spherical_moment_3rd}, \link{spherical_moment_4th}
}
