% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symm_mat.R
\name{symm_mat}
\alias{symm_mat}
\title{Symmetric matrix transform}
\usage{
symm_mat(a, m)
}
\arguments{
\item{a}{vector of the upper matrix.}

\item{m}{dimension parameter of symmetric matrix.}
}
\description{
Inverse of half-vectorisation function \code{vech}. In other words,
it creates a symmetric matrix using entries from the upper triangular matrix.
}
\examples{
symm_mat(1:3)
##      [,1] [,2]
## [1,]    1    2
## [2,]    2    3

symm_mat(m=2)
##      [,1] [,2]
## [1,]    1    2
## [2,]    2    3
}
\seealso{
\link{vech}
}
