% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vech.R
\name{vech}
\alias{vech}
\title{Half-vectorization}
\usage{
vech(A)
}
\arguments{
\item{A}{symmetric matrix.}
}
\value{
vector with elements from the lower triangular matrix.
}
\description{
Vectorising the lower triangular part of a symmetric matrix
}
\examples{
A <- symm_mat(m=2)
##      [,1] [,2]
## [1,]    1    2
## [2,]    2    3

vech(A)
## [1] 1 2 3
}
\seealso{
\link{symm_mat}
}
